Правильное чтение при загрузке, выгрузке. Папка Unsorted

+++ Features/MapUtils.cs
@@ -1,3 +1,4 @@
+using System.Drawing.Drawing2D;
 using Exiled.API.Features;
 using Exiled.Loader;
 using PlayerRoles;
@@ -17,8 +18,7 @@ public static class MapUtils
     public static MapSchematic UntitledMap => LoadedMaps.GetOrAdd(UntitledMapName, () => new(UntitledMapName));
 
     public static Dictionary<string, MapSchematic> LoadedMaps { get; private set; } = [];
-
-    // TODO: Нужно как-то правильно сохранять, ровно как и правильно грузить
+    
     public static void SaveMap(string mapName)
     {
         if (mapName == UntitledMapName)
@@ -31,7 +31,14 @@ public static class MapUtils
         else // Map isn't loaded and map file doesn't exist
             map = new MapSchematic(mapName).Merge(UntitledMap);
         
-        string path = Path.Combine(ProjectMER.MapsDir, $"{mapName}.yml");
+        string path = Path.Combine(ProjectMER.MapsDir, "UNSORTED");
+        if (!Directory.Exists(path))
+        {
+            Log.Warn($"Map saving is not enabled. To enable it, create directory named 'UNSORTED' in your maps directory");
+            return;
+        }
+        
+        path = Path.Combine(path, $"{map.Name}.yml");
         File.WriteAllText(path, Loader.Serializer.Serialize(map));
         map.IsDirty = false;
 
@@ -81,8 +88,10 @@ public static class MapUtils
         string? foundPath = null;
         foreach (var mapFile in FileExtensions.GetAllMaps())
         {
-            if (Path.GetFileName(mapFile) != $"{mapName}.yml")
+            string name = Path.GetFileName(mapFile);
+            if (name != $"{mapName}.yml")
             {
+                Log.Info($"Найдено: {name}, нужно найти: {mapName}.yml");
                 continue;
             }
 
@@ -90,7 +99,6 @@ public static class MapUtils
             break;
         }
         
-        // TODO: Почему-то не находит карту
         if (foundPath == null)
         {
             string error = $"Failed to load map data: File {mapName}.yml does not exist!";
@@ -126,12 +134,35 @@ public static class MapUtils
             return false;
         }
     }
-
-    // TODO: Проделать те же махинации по поиску и взятию схематика.
+    
     public static SchematicObjectDataList GetSchematicDataByName(string schematicName)
     {
         SchematicObjectDataList data;
-        string schematicDirPath = Path.Combine(ProjectMER.SchematicsDir, schematicName);
+        string? schematicDirPath = null;
+        
+        foreach (string ownerDirectory in Directory.GetDirectories(ProjectMER.SchematicsDir))
+        {
+            foreach (var schematicDirectory in Directory.GetDirectories(ownerDirectory))
+            {
+                string name = Path.GetFileName(schematicDirectory);
+                if (name is null || name != schematicName)
+                    continue;
+
+                Log.Info("Нашлось.");
+                schematicDirPath = schematicDirectory;
+                break;   
+            }
+            
+            if (schematicDirPath != null)
+                break;
+        }
+
+        if (schematicDirPath == null)
+        {
+            Log.Info($"Схематик '{schematicName}' не найден");
+            return null;
+        }
+        
         string schematicJsonPath = Path.Combine(schematicDirPath, $"{schematicName}.json");
         string misplacedSchematicJsonPath = schematicDirPath + ".json";
